Post microservice: move restTemplate logic to controller from service
