+++ post-service/src/main/java/com/hexagon/postservice/controller/PostController.java
@@ -1,7 +1,9 @@
 package com.hexagon.postservice.controller;
 
+import com.hexagon.authservice.dto.UserResponse;
 import com.hexagon.postservice.dto.PostRequest;
 import com.hexagon.postservice.dto.PostResponse;
+import com.hexagon.postservice.entity.Post;
 import com.hexagon.postservice.service.PostService;
 import java.util.List;
 import java.util.Optional;
@@ -10,6 +12,7 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 @CrossOrigin("*")
 @RequestMapping(value = "/post")
@@ -19,24 +22,35 @@ public class PostController {
 
   @Autowired private PostService postService;
 
+  @Autowired private RestTemplate restTemplate;
+
+  private static final String authServiceUrl = "http://AUTH-SERVICE/auth/user";
+
   @PostMapping("/addPost")
   public PostResponse addPost(@RequestBody PostRequest postRequest, @RequestHeader String token) {
-    return postService.addPost(postRequest, token);
+    String getAuthorReqUrl = authServiceUrl + "?token=" + token;
+    UserResponse author = restTemplate.getForObject(getAuthorReqUrl, UserResponse.class);
+    Post post = new Post(postRequest, author.getId());
+
+    return getPostResponse(postService.addPost(post));
   }
 
   @GetMapping
   public List<PostResponse> getPosts() {
-    return postService.getPosts();
+    return postService.getPosts().stream().map(this::getPostResponse).toList();
   }
 
   @GetMapping("/{id}")
   public ResponseEntity<?> getPostById(@PathVariable int id) {
-    Optional<PostResponse> postResponse = postService.getPostById(id);
+    Optional<PostResponse> postResponse = postService.getPostById(id).map(this::getPostResponse);
+    return ResponseEntity.of(postResponse);
+  }
 
-    if (postResponse.isPresent()) {
-      return ResponseEntity.ok(postResponse.get());
-    } else {
-      return ResponseEntity.notFound().build();
-    }
+  private UserResponse getPostAuthor(Post post) {
+    return restTemplate.getForObject(authServiceUrl + "/" + post.getUserId(), UserResponse.class);
+  }
+
+  private PostResponse getPostResponse(Post post) {
+    return new PostResponse(post, getPostAuthor(post));
   }
 }
